Minor miscellaneous tweaks
Changes the sound of the zero-point laser to use the, what I believe to be unused, emitter2.ogg.

Fixes a missing tile on the Syndicate Mothership, which even though is not necessary the fact it's missing bugs me greatly.

Increases the default layer of blast doors (poddoors) to 3.0, making them more visible when on cluttered tiles in the editor. This is set back to standard for actual gameplay.

Fixed an incorrectly typed firelock and added a missing emergency shutter in the Custodial Closet.