Controller/event/misc fixes and lighting tweaks
Disables the old 'events' system and onlines the new 'events' process to replace it.

Adds a missing option to the admin panel to spawn an electrical storm.

Adds some more fixes to /atom and /atom/movable to try and make lighting a bit more stable.

Fixes multiple miscellaneous light sources not emitting correctly.

Increases brightness and range of lights to match what they previously were before the new system was added.

Updates variable modification to use 'light_range' instead of 'luminosity' values.

Fixes some fire-related lighting code that was wonky.

Attempts to fix a runtime in beams.dm.

Reduces docking time modifier of the arrivals shuttle to a maximum of +30 instead of +45.