Map lighting pass part I - shuttles
Adds lighting to the following shuttles:
- Arrivals
- Engineering
- Mining
- Research
- Centcom Emergency
- Administration
- Transport
- Spec Ops
- Syndicate Elite
- Syndicate Mothership
- Vox Skipjack
- Supply

Adds lighting to escape pods one, two, three and five.

Edits lighting on the Emergency Shuttle to better fit the new system.

Fixes a missing table on the Centcom Emergency Shuttle.

Disables dynamic lighting use in the Admin Room, Centcom, Syndicate Forward Operating Base, Syndicate Control Room, Syndicate Elite Squad, Wizard's Den, Centcom Solitary Confinement and the Thunderdome.

Edits the sprite for /area/space to make it more visible in the editor.

Hopefully this gives an idea of the aesthetic I want to have in terms of the retro "dark and gloomy" look.