Ports minor fixes
Ports UristMcKerman's fix for a bug with the TEG not updating gas correctly.

Ports mwerezak's fix for a runtime in null.virus2, and for lightbulb boxes not being able to have items reinserted.

Ports mwerezak's fixes for several consoles being unbuildable/undeconstructable.

Additionally fixes a typo in the output messages for curer.dm.

And finally does some file formatting cleanup in the mining module.