Span classes and minor tweaks part I
Adds span classes to the mining module and ZAS code.

Adds mythril coins to the list of allowed items for a cash bag.

Adds a mythril coin to the vault's coin storage, since there was already an adamantium one there.

First commit from new PC, everything seems to be okay so far.