Radio and supply changes
The Quartermaster can now access the command radio channel.

The Head of Personnel can now access the supply radio channel.

Rearranged radio frequencies into static defines so that nothing gets mixed up later on.

Re-added the mining radio frequency and associated headsets, with the mining channel accessible to:
- Shaft Miner
- Mining Foreman
- Quartermaster
- Head of Personnel
- Emergency Response Team
- Deathsquad

Updates the Mining Foreman to have their own unique headset and encryption key, with access to both supply and mining channels, alongside common.

Updates Shaft Miners to spawn with their own new headset and encryption key, with access to only the mining and common channels.

Cleans up some job-related on-spawn text to fix capitalization.

Fixes some text issues in the descriptions of various headsets.

Adds the Gardener alt-title to the Botanist.

Adds the Reporter alt-title to the Librarian.

Adds the Custodial Specialist and Sanitation Technician alt-titles to the Janitor.

Adds the Drone alt-title to Cyborgs.

Does some more cleanup of job-related code.

Updates the changelog, need to get into the habit of that.