Fixes ProcessScheduler exception handling
Fixes a runtime error in the ProcessScheduler's exception handling proc.

This is a funny one, I never had an error thrown from within a process before, until I manually introduced one for debugging purposes. Therefore I never knew that the code to handle it, itself had a runtime in it!