The Mappening I & II
The mappening is happening.

The Mappening I: The Spacious Medical Treatment Centre.

Slightly remaps the Medical Treatment Centre to have more spaced out cryotubes, including a revamped cooling loop with a meter to measure gas pressure.

Swaps the positions of the NanoMed and disposal unit situated between Chemistry and Emergency Response Auxiliary Storage.

Moves a light fixture in the Medbay Hallway to accommodate the movement of walls in the Medical Treatment Centre.

The Mappening II: Return of the Vacant Office.

Replaces the Disused Security Dormitories with the long-awaited faithful recreation of the iconic Vacant Office.

Changelog update included.