The Mappening III: Chapels and Cryogenics
Fixes the red and yellow chapel floor decals so that they now have the correct sprites for them.

Adds a new "darkchapel" floor tile sprite, which is intended for use underneath the chapel floor decals to maintain the same appearance as previous.

Converts all chapel floor tiles on station Z-levels 1 and 4, as well as the Academy and Station Collision away missions to use the new floor decal overlay system.

Moves the meter connected to the cryogenics tube cooling loop one tile to the right, because MadMurky said it satisfied their OCD that way.

Replaces the Research Server Room cooling loop's air canister with a nitrogen canister, actually this time.

Changelog update included.