Moves code/stylesheet.dm to interface/stylesheet.dm
Moves this rather important file to a better place as I don't see a need for it to be cluttering up the code folder.

It's also only really related to the interface instead of the code, so why wasn't it put there originally?

Adds a missing / to the start of the file.

No changelog since backend.