Updates __DEFINES/jobs.dm, CMO and RD lockers
Repaths the CMO and RD lockers so that they use lowercase cmo and rd to avoid variable conflicts.

Converts the /var/const department and job bitflags to #defines, prefixed with DEPARTMENT_ and JOB_ respectively.

Updates the mapped in instances of CMO and RD lockers to reflect path changes.

No changelog since this should be invisible.