Updates SpacemanDMM debugging hooks
Adds /datum/debugger, initialised during global_init so that the debugger runs from the very word go.

Moves the global.debugger.shutdown_debugger() call to be the final statement before the return in /world/Del(), so it's the very last thing to be stopped.

This means we can now debug global initialisation hooks and shutdown hooks!

No changelog as it's backend.