Fixes XGM/ZAS filename capitalisation
Uppercases the __DEFINES/atmospherics/ZAS.dm and __DEFINES/atmospherics/XGM.dm filenames for consistency.

Uppercases the XGM module's folder name for more consistency.

It turns out XGM is an acronym, we don't know what it means (eXtensible Gas Mixtures?), but it's an acronym! Special thanks to Chinsky for confirming this for me.

No changelog update, it's just backend OCD prevention.