Telecoms spelling standardisation
Standardises Telecomms -> Telecoms, telecomms -> telecoms and tcomms to tcoms to match the correct spelling.

Renames a few folders and files that had the incorrect spelling.

Adjusts telecoms satellite area naming to use the full 'Telecommunications X' format as per the original intention.

This is basically a large find and replace with some manual effort thrown in.

Changelog update included.