Updates many shuttle windows to use spawners
Replaces applicable mapped-in windows on shuttles with the new spawner objects. A full-tile window is still manually placed over the top however.

This doesn't apply to windows that would link to windows they shouldn't be linked with.

This also cuts another millisecond or so off of atmospherics geometry initialisation.

Changelog update included.