Cleans up the job controller
Moves the 'You are the...' spawn text to an earlier point in the equip_rank() proc so that it's shown to cyborgs.

Sets the idtype var on the AI and cyborg jobs to null.

Updates PDA and ID setup to use the new helper functions.

Removes some unnecessary return values.

Updates setup_occupations() to use the SUBTYPESOF macro.

Changelog update included.