Updates occupation choices UI
Converts the occupation choices panel to use /datum/browser and the newer "fake nanoUI" browser style.

Adjusts the formatting of the occupation choices panel so it looks less strange.

Adjusts the layout of the occupation choices panel for consistency with other panels.

Makes the names of jobs with alt-titles clickable to select them instead of adding a wonky second line of text.

Adjusts the wording of the "alternate"/"fallback" preference option.

I never thought I was going to be able to do this one, but after spending all night on it, here we are!

Changelog update included!