Cleans up the XGM module
Updates a lot of comment formatting to use the new style.

Fixes many code formatting and style issues.

Removes a redundant second moles > 0 check from adjust_gas_temp().

Converts a boolean check to an isnull().

No changelog update as it's backend, and performance improvements are already on there if this does any.