Optimises exosuit equipment exclusions
Adds an excluded equipment list to all exosuits to make managing what can and cannot be equipped on various exosuit types.

Refactors the exosuit equipment can_attach() proc to be less spaghetti and use the new exclusion list.

Makes the hydraulic clamp and kill clamp fittable on the hoverpod as well as the ripley.

Changelog update included!