Generalises mecha leg actuator overload
Generalises the mecha leg actuator overload functionality away from /obj/mecha/combat/gygax and onto /obj/mecha.

Adds /obj/mecha/var/overload_capable as a toggle for whether or not a specific mecha can activate the leg actuator overload.

Adds leg overload functionality to the new serenity mecha as it's based on the gygax. This technically isn't an actual addition since it was always intended to have that, I just had to split up the commits.

No changelog update as it's backend.