Adds basic 516 compatibility
Merges the various SIGN() macros to use the global /proc/sign().

Adds a condition to switch to the new BYOND built-in sign() proc in 516 and upwards.

Renames a variable in /proc/call_hook() from caller to called_hook to avoid future variable naming collisions.

No changelog update as it's backend.