Tweaks centcom and syndicate mech loadouts
Replaces the odysseus in the centcom special operations area with a dark odysseus.

Replaces one of the maulers in the syndicate elite squad mech room with a eurymachus.

Fixes the naming of the syndicate elite squad mech room airlock.

Updates the layouts of the exosuit charging ports in centcom special operations and the syndicate elite squad mech room for consistency.

Changelog update included!