Adds the Dark Phazon exosuit
Adds the dark phazon exosuit, an admin-only variant of the phazon which functions as its equivalent to the dark gygax while also fulfilling the role of the original non-constructable phazon.

Updates phazon phasing icon state to use /obj/mecha/var/initial_icon so it's no longer hardcoded for subtypes.

The custom phasing effect attempts to replicate the original using three frames with a square blur of 0.1 -> 1.0 -> 2.0 respectively. It's not quite there but it's the closest I could get!

Changelog update included!