Re-adds custom exosuit on-entry directions
Re-adds the ability for exosuits to assign a custom direction that they face when entered or exited.

Updates the gygax, dark gygax, durand, archambeau, phazon and dark phazon exosuits so they correctly point north when entered or exited. It turns out that on closer examination of the sprites, you're climbing into these from the back!

Changelog update included!