Overhauls mecha construction costs
Reduces the steel costs of the vast majority of exosuit parts by 40% rounded to the nearest multiple of 2000.

Reduces the plasma cost of phazon armour from 30000 to 18000.

Adds a 6000 plasma cost to the odysseus and serenity carapace parts.

Increases the steel cost of the odysseus head from 2000 to 6000.

This is a long overdue part II of the material stack amount standardisation. Where I cut all of the material stacks from 3750 to 2000 per sheet, that's a reduction of 46.6667%. What I forgot to do was decrease many of the costs accordingly!

Changelog update included!