Overhauls mecha construction costs part II
Increases the default glass cost of exosuit chassis from 5000 to 6000.

Reduces the plasma and uranium costs of archambeau armour from 30000 to 18000 of each.

Increases the silver cost of durand heads , arms and legs from 5000 to 6000.

Increases the glass cost of phazon, honk and reticence heads from 5000 to 6000.

Increases the bananium cost of honk and reticence heads, arms and legs from 5000 to 6000.

Increases the glass cost of ripley torsos from 15000 to 16000.

Fixes some changelog oopsies.

Changelog update included!