Adds the APLU "Sindy", tweaks exosuits
Adds a syndicate version of the ripley exosuit which is slightly faster and can equip weapons. The name is a double barrel reference to the shortform "syndie" but also british doll sindy created as a competitor to barbie.

Replaces one of the three maulers in the syndicate elite squad mech room with a syndicate ripley.

Tweaks marauder, seraph and mauler exosuits so they come with hyper-capacity power cells as standard.

Fixes some extraneous white pixels on the ripley exosuit's broken sprite.

Rearranges the default equipment configuration of the eurymachus exosuit so that the taser comes before the medical equipment.

Changelog update included!