Ports Sandstorm Station's screentips and status bar
Removes the default bulky status bar from the bottom of the byond window to free up more space.

Adds a new small tg-style status bar to the bottom left corner of the byond window.

Adds screentips which display the name of the atom the mouse cursor is hovering over at the top of the screen.

Adds ATOM_FLAG_NO_SCREENTIP for future use by atoms that don't want to display a screentip on hover.

Adds preference options for enabling or disabling screentips and setting their text colour.

This is basically a port from Sandstorm Station coupled with a few of my own tweaks. Absolutely massive thanks to SandPoot for helping me get this working after several attempts.

Changelog update included!