Begins refactoring autolathes, tweaks mint colours
Significantly cleans up autolathe code in preparation for the design datum and material container refactors.

Standardises autolathe storage capacity to be the same for all materials.

Standardises various pieces of autolathe feedback text.

Updates the autolathe UI to be more consistent with other material-using machinery.

Updates the autolathe description to refer to steel instead of metal.

Slightly improves the performance of autolathes.

Adjusts the coin press colour of adamantine to make it less similar to that of iron.

Adds a colour code to /decl/material/glass for use by autolathes.

Fixes a changelog oopsie.

Changelog update included!