Adds robot lights UI button, removes verb
Adds a UI element to the robot hud for toggling of their integrated lights.

Removes the toggle lights robot verb as it's no longer necessary.

Tweaks the screen1_robot.dmi's "inv1" icon state to make the new element fit on the hud.

Moves var/integrated_light_power to the robot model instead of the mob type.

Increases the brightness of mining robot lights to match that of a mining lantern.

Changelog update included!