Updates the defunct end of the world event
Removes the defunct 'end of the world' event as we all actually survived the 21st december 2012 (I was there!).

Updates the configuration name for the 'end of the world' event to the 'end of the world anniversary' in celebration of what a cultural phenomenon it was.

No changelog update as it's backend!