Remaps the telecommunications satellite
Converts the atmos piping to use mains pipes.

Adds air scrubbers to almost every room.

Replaces the four solar trackers with a single one directly outside of the main airlock.

Moves the solar control computer and SMES to the main teleporter room.

Fixes many wires travelling under walls and windows.

Rearranges many interior decorations to reflect the fact that a lot of stuff moved.

Updates the solar control computer and SMES so that they start activated.

Makes the two window grilles in the teleporter room electrified as I (think?) they were intended to be originally.

Adds a new pre-activated variant of the solar control computer for mapping purposes.

This new version of the satellite is inspired even more by the Polaris version, so thanks to them for their excellent work.

Changelog update included!