Adds exosuit hardened ablative EMP insulation
Adds a hardened variant of the exosuit ablative EMP insulation with a slightly stronger EMP severity modifier of 0.75 instead of 0.85.

Updates the loadouts of all pre-equipped syndicate and centcom exosuits to feature the new hardened ablative EMP insulation.

Increases the maximum equipment slots of the marauder, seraph, mauler and dark gygax by 1. This is in practice not a change as the new slot gets filled by unremovable EMP insulation anyway.

Changelog update included!