Remaps virology
Redesigns virology to be more modern and space-efficient.

Removes the monkey pen in favour of monkey cube boxes.

Moves various items around so that things are more spread out and less stacked in large piles.

Shortens the medical bay aft corridor to reduce some wasted space.

Special thanks to Noize for his input and contributions on this!

Changelog update included!