Adds the Brigand exosuit, fixes some issues
Fixes durand head fabrication not requiring glass.

Fixes a runtime when completing exosuit construction.

Adds the brigand exosuit, designed and sprited by MetalClone with some additions by me, as a prototype version of the marauder. Designed to bridge the gap between the durand and marauder exosuits while still having a unique gimmick, it features both the durand's defence mode and the marauder's zoom mode along with its own special mounted emitter weapon.

Adds research designs for the new brigand exosuit components.

Adds the NT-EM mounted emitter exosuit weapon for the brigand.

Special thanks to MetalClone for his sprite and design work which is the whole reason this thing exists, and for letting me go wild with it to fill in the gaps!

Changelog update included!