Adds the Fiend exosuit
Adds the fiend exosuit as the syndicate counterpart to the seraph. This is basically the seraph's version of the mauler.

Adds a fiend to the syndicate elite squad mech room to replace one of the two maulers.

Special thanks to Justice12354 for inspiring me to sprite this!

Changelog update included!