Adds the Thales exosuit part I
Adds the thales exosuit. Originally concepted and sprited by MetalClone with some edits by me.

Adds a unique laser pulse sound created by DrDiasyl for the axis energy projector.

The actual exosuit sprites themselves aren't included in this as they're still WIP.

This ballooned out of scope so much that I'm going to have to make it a PR instead of a regular commit!